\name{implants}       \docType{data}       \alias{implants}       \title{Fetal Death in Mice}       \description{         Data on the number of fetal deaths arising from dominant lethal testing in mice.       }       \usage{data(implants)}       \format{A data frame containing the number of dead and survived implants from 523 mice.}       \source{Zhou H and Lange K (2010). MM Algorithms for Some Discrete Multivariate Distributions. \emph{Journal of Computational and Graphical Statistics}. 19 (3) 645-665. Supplementary material.}
	\references{Haseman JK and Soares ER (1976). The Distribution of Fetal Death in Control Mice and Its Implications on Statistical Tests for Dominant Lethal Effects, \emph{Mutation Research/Fundamental and Molecular Mechanisms of Mutagenesis}. 41, 277-288.}       \keyword{datasets}