\name{parties}       \docType{data}       \alias{parties}       \title{Political Parties}       \description{         Political parties of members of the U.S. House of Representatives, 2005.       }       \usage{data(votes)}       \format{A vector of integers representing the political parties, with 0, 1, and 2 corresponding to Republicans, Democrats, and Independents, respectively.}       \source{Diaconis PD, Goel S, and Holmes S (2008). Horseshoes in Multidimensional Scaling and Local Kernel Methods. \emph{Annals of Applied Statistics}. 2 (3) 777-807. Supplementary material.}       \keyword{datasets}