\name{rats}       \docType{data}       \alias{rats}       \title{Population Growth of Rats}       \description{         Longitudinal measurements of the weights of rats in control and treatment groups.       }       \usage{data(rats)}       \format{A data frame containing weights from 60 rats divided into two groups, with each rat measured at 5 time points.}       \source{Gelfand AE, Hills SE, Racine-Poon A, and Smith AFM (1990). Illustration of Bayesian inference in normal data models using Gibbs sampling. \emph{Journal of the American Statistical Association}. 85, 972-985.}       \keyword{datasets}