\name{votes}       \docType{data}       \alias{votes}       \title{Roll Call Votes}       \description{         Roll call votes from the U.S. House of Representatives, 2005.       }       \usage{data(votes)}       \format{A 401-by-669 matrix whose (\eqn{i}{i},\eqn{j}{j})-entry corresponds to the vote of the \eqn{i}{i}th representative on the \eqn{j}{j}th roll call. Possible votes are "yea", "nay", or "not voting", which are represented by 1/2, -1/2, and 0, respectively.}       \source{Diaconis PD, Goel S, and Holmes S (2008). Horseshoes in Multidimensional Scaling and Local Kernel Methods. \emph{Annals of Applied Statistics}. 2 (3) 777-807. Supplementary material.}       \keyword{datasets}